+++ lib/main-rn.ts
@@ -1,0 +1,4 @@
+import { nanoqueryFactory } from "./factory";
+import { reactNativeCompat } from "./platforms/react-native";
+
+export const nanoquery = nanoqueryFactory(reactNativeCompat);
