+++ lib/platforms/type.ts
@@ -1,0 +1,9 @@
+type IsAppVisible = () => boolean;
+type VisibilityChangeSubscribe = (cb: () => void) => void;
+type ReconnectChangeSubscribe = (cb: () => void) => void;
+
+export type PlatformCompat = [
+  IsAppVisible,
+  VisibilityChangeSubscribe,
+  ReconnectChangeSubscribe
+];
